Issue #1847406 by juampy: Move account listing to the module settings.
Added added_by_uid field to twitter_account table in order to list
Twitter accounts added by Drupal users at the Twitter Accounts listing.
This is needed because when we pull mentions the twitter_account table
gets populated automatically with the Twitter accounts that mention it
and we do not want these to be listed in the admin interface.